requirements.txt
python-docx
requests
appdirs==1.4.4
certifi==2020.6.20
click==7.1.2
distlib==0.3.1
elasticsearch==7.9.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Ext==0.1
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
